% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abalone_docs.R
\docType{data}
\name{abalone}
\alias{abalone}
\title{Abalone Data Set}
\format{A data frame with 4177 observations on the following 9 variables.
\itemize{
\item \code{sex}: Factor
\itemize{
\item \code{M} (Male), \code{F} (Female), and \code{I} (Infant)
}
\item \code{length}: Numeric
\itemize{
\item Longest shell measurement (mm)
}
\item \code{diameter}: Numeric
\itemize{
\item Perpendicular to length (mm)
}
\item \code{height}: Numeric
\itemize{
\item With meat in shell (mm)
}
\item \code{whole_weight}: Numeric
\itemize{
\item Whole abalone weight (grams)
}
\item \code{shucked_weight}: Numeric
\itemize{
\item Weight of meat (grams)
}
\item \code{viscera_weight}: Numeric
\itemize{
\item Gut weight after bleeding (grams)
}
\item \code{shell_weight}: Numeric
\itemize{
\item Shell weight after being dried (grams)
}
\item \code{rings}: Integer
\itemize{
\item Adding 1.5 gives the age in years
}
}}
\source{
Marine Resources Division
Marine Research Laboratories - Taroona
Department of Primary Industry and Fisheries, Tasmania
GPO Box 619F, Hobart, Tasmania 7001, Australia
(contact: Warwick Nash +61 02 277277, wnash '@' dpi.tas.gov.au)
}
\usage{
abalone
}
\description{
Predicting the age of abalone from physical measurements. The age of abalone
is determined by cutting the shell through the cone, staining it, and
counting the number of rings through a microscope -- a boring and
time-consuming task. Other measurements, which are easier to obtain, are
used to predict the age. Further information, such as weather patterns and
location (hence food availability) may be required to solve the problem.
}
\references{
Warwick J Nash, Tracy L Sellers, Simon R Talbot, Andrew J Cawthorn and Wes B Ford (1994)
"The Population Biology of Abalone (\emph{Haliotis} species) in Tasmania. I. Blacklip Abalone (\emph{H. rubra}) from the North Coast and Islands of Bass Strait",
Sea Fisheries Division, Technical Report No. 48 (ISSN 1034-3288)
}
\keyword{datasets}
