% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_docs.R
\docType{data}
\name{adult}
\alias{adult}
\title{Adult Data Set}
\format{
A data frame with 32561 observations on the following 15 variables.
\itemize{
\item \code{age}: Integer
\itemize{
\item Number of years alive
}
\item \code{workclass}: Factor
\itemize{
\item Private, Self-emp-not-inc, Self-emp-inc, Federal-gov, Local-gov,
State-gov, Without-pay, Never-worked.
}
\item \code{fnlwgt}: Numeric
\itemize{
\item The variable represents the Final Weight, which is more so a sampling weight.
See the names file listed in references for more details.
}
\item \code{education}: Factor
\itemize{
\item Highest level of education attained
\item Bachelors, Some-college, 11th, HS-grad, Prof-school, Assoc-acdm,
Assoc-voc, 9th, 7th-8th, 12th, Masters, 1st-4th, 10th, Doctorate,
5th-6th, Preschool.
}
\item \code{education_num}: Numeric
\itemize{
\item Number of years of education
}
\item \code{marital_status}: Factor
\itemize{
\item Married-civ-spouse, Divorced, Never-married, Separated, Widowed,
Married-spouse-absent, Married-AF-spouse
}
\item \code{occupation}: Factor
\itemize{
\item Tech-support, Craft-repair, Other-service, Sales, Exec-managerial,
Prof-specialty, Handlers-cleaners, Machine-op-inspct, Adm-clerical,
Farming-fishing, Transport-moving, Priv-house-serv, Protective-serv,
Armed-Forces.
}
\item \code{relationship}: Factor
\itemize{
\item Wife, Own-child, Husband, Not-in-family, Other-relative, Unmarried.
}
\item \code{race}: Factor
\itemize{
\item White, Asian-Pac-Islander, Amer-Indian-Eskimo, Other, Black.
}
\item \code{sex}: Factor
\itemize{
\item Female, Male
}
\item \code{capital_gain}: Integer
\itemize{
\item Income from investment sources, apart from wages/salary
}
\item \code{capital_loss}: Integer
\itemize{
\item Losses from investment sources, apart from wages/salary
}
\item \code{hours_per_week}: Integer
\itemize{
\item Amount of hours worked per week
}
\item \code{native_country}: Factor
\itemize{
\item Country of origin
\item United-States, Cambodia, England, Puerto-Rico, Canada, Germany,
Outlying-US(Guam-USVI-etc), India, Japan, Greece, South, China, Cuba,
Iran, Honduras, Philippines, Italy, Poland, Jamaica, Vietnam, Mexico,
Portugal, Ireland, France, Dominican-Republic, Laos, Ecuador, Taiwan,
Haiti, Columbia, Hungary, Guatemala, Nicaragua, Scotland, Thailand,
Yugoslavia, El-Salvador, Trinadad&Tobago, Peru, Hong, Holand-Netherlands.
}
\item \code{income}: Factor
\itemize{
\item Whether the income greater than $50,000 or not.
\item <=50K, >50K
}
}
}
\source{
Ronny Kohavi and Barry Becker
Data Mining and Visualization
Silicon Graphics.
e-mail: ronnyk '@' live.com for questions.
}
\usage{
adult
}
\description{
Extraction was done by Barry Becker from the 1994 Census database.
A set of reasonably clean records was extracted using the following
conditions: ((AAGE > 16) && (AGI > 100) && (AFNLWGT > 1) && (HRSWK > 0))
}
\details{
Prediction task is to determine whether a person makes over 50K a year.
}
\references{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/adult/}

\url{http://archive.ics.uci.edu/ml/machine-learning-databases/adult/adult.names}

\url{https://archive.ics.uci.edu/ml/datasets/adult}
}
\keyword{datasets}
