% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoimports_docs.R
\docType{data}
\name{autoimports}
\alias{autoimports}
\title{Autoimports Dataset}
\format{A data frame with 205 observations on the following 26 variables.
\itemize{
\item \code{symboling}:
\itemize{
\item -3, -2, -1, 0, 1, 2, 3.
}
\item \code{normalized-losses}:
\itemize{
\item continuous from 65 to 256.
}
\item \code{make}:
\itemize{
\item alfa-romero, audi, bmw, chevrolet, dodge, honda, isuzu,
jaguar, mazda, mercedes-benz, mercury, mitsubishi, nissan,
peugot, plymouth, porsche, renault, saab, subaru, toyota, volkswagen,
volvo
}
\item \code{fuel-type}:
\itemize{
\item diesel, gas.
}
\item \code{aspiration}:
\itemize{
\item std, turbo.
}
\item \code{num-of-doors}:
\itemize{
\item four, two.
}
\item \code{body-style}:
\itemize{
\item hardtop, wagon, sedan, hatchback, convertible.
}
\item \code{drive-wheels}:
\itemize{
\item 4wd, fwd, rwd.
}
\item \code{engine-location}:
\itemize{
\item front, rear.
}
\item \code{wheel-base}:
\itemize{
\item continuous from 86.6 120.9.
}
\item \code{length}:
\itemize{
\item continuous from 141.1 to 208.1.
}
\item \code{width}:
\itemize{
\item continuous from 60.3 to 72.3.
}
\item \code{height}:
\itemize{
\item continuous from 47.8 to 59.8.
}
\item \code{curb-weight}:
\itemize{
\item continuous from 1488 to 4066.
}
\item \code{engine-type}:
\itemize{
\item dohc, dohcv, l, ohc, ohcf, ohcv, rotor.
}
\item \code{num-of-cylinders}:
\itemize{
\item eight, five, four, six, three, twelve, two.
}
\item \code{engine-size}:
\itemize{
\item continuous from 61 to 326.
}
\item \code{fuel-system}:
\itemize{
\item 1bbl, 2bbl, 4bbl, idi, mfi, mpfi, spdi, spfi.
}
\item \code{bore}:
\itemize{
\item continuous from 2.54 to 3.94.
}
\item \code{stroke}:
\itemize{
\item continuous from 2.07 to 4.17.
}
\item \code{compression-ratio}:
\itemize{
\item continuous from 7 to 23.
}
\item \code{horsepower}:
\itemize{
\item continuous from 48 to 288.
}
\item \code{peak-rpm}:
\itemize{
\item continuous from 4150 to 6600.
}
\item \code{city-mpg}:
\itemize{
\item continuous from 13 to 49.
}
\item \code{highway-mpg}:
\itemize{
\item continuous from 16 to 54.
}
\item \code{price}:
\itemize{
\item continuous from 5118 to 45400.
}
}}
\source{
Donor: Jeffrey C. Schlimmer (Jeffrey.Schlimmer@a.gp.cs.cmu.edu)

1985 Model Import Car and Truck Specifications, 1985 Ward's Automotive Yearbook.

Personal Auto Manuals, Insurance Services Office, 160 Water Street, New York, NY 10038

Insurance Collision Report, Insurance Institute for Highway Safety, Watergate 600, Washington, DC 20037
}
\usage{
autoimports
}
\description{
This data set consists of three types of entities:
(a) the specification of an auto in terms of various characteristics,
(b) its assigned insurance risk rating,
(c) its normalized losses in use as compared to other cars.
The second rating corresponds to the degree to which the auto is more risky
than its price indicates. Cars are initially assigned a risk factor symbol
associated with its price. Then, if it is more risky (or less), this symbol
is adjusted by moving it up (or down) the scale. Actuarians call this
process "symboling". A value of +3 indicates that the auto is risky, -3
that it is probably pretty safe. The third factor is the relative average
loss payment per insured vehicle year. This value is normalized for all autos
within a particular size classification (two-door small, station wagons,
sports/speciality, etc...), and represents the average loss per car per year.
}
\references{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/autos/imports-85.data}
\url{https://archive.ics.uci.edu/ml/datasets/Automobile}
}
\keyword{datasets}
