% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_datasets.R
\docType{data}
\name{autompg}
\alias{autompg}
\title{Autompg Data Set}
\format{A data frame with 6497 observations (1599 Red and 4898 White) on the following 12 variables.
\itemize{
\item \code{mpg}: continuous
\item \code{cylinders}:     multi-valued discrete
\item \code{displacement}:  continuous
\item \code{horsepower}:    continuous
\item \code{weight}:        continuous
\item \code{acceleration}:  continuous
\item \code{model_year}:    multi-valued discrete
\item \code{origin}:        multi-valued discrete
\item \code{car_name}:      string (unique for each instance)
}}
\source{
This dataset was taken from the StatLib library which is
maintained at Carnegie Mellon University. The dataset was
used in the 1983 American Statistical Association Exposition.
}
\usage{
autompg
}
\description{
This dataset is a slightly modified version of the dataset provided in
the StatLib library.  In line with the use by Ross Quinlan (1993) in
predicting the attribute "mpg", 8 of the original instances were removed
because they had unknown values for the "mpg" attribute.
}
\references{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.names}
\url{https://archive.ics.uci.edu/ml/datasets/auto+mpg}
}
\keyword{datasets}
