% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike_sharing_daily_docs.R
\docType{data}
\name{bike_sharing_daily}
\alias{bike_sharing_daily}
\title{Bike Sharing (Daily) Data Set}
\format{
A data frame with 731 observations on the following 16 variables.
\itemize{
\item \code{instant}: Record index
\item \code{dteday}:     Date
\item \code{season}:
\itemize{
\item 1: Spring
\item 2: Summer
\item 3: Fall
\item 4: Winter
}
\item \code{yr}:
\itemize{
\item 0: 2011
\item 1: 2012
}
\item \code{mnth}:
\itemize{
\item 1: Jan
\item 2: Feb
\item 3: Mar
\item 4: Apr
\item 5: May
\item 6: Jun
\item 7: Jul
\item 8: Aug
\item 9: Sep
\item 10: Oct
\item 11: Nov
\item 12: Dec
}
\item \code{hr}:
\itemize{
\item 0: 12 AM
\item 1:  1 AM
\item 2:  2 AM
\item 3:  3 AM
\item 4:  4 AM
\item 5:  5 AM
\item 6:  6 AM
\item 7:  7 AM
\item 8:  8 AM
\item 9:  9 AM
\item 10: 10 AM
\item 11: 11 AM
\item 12: 12 PM
\item 13:  1 PM
\item 14:  2 PM
\item 15:  3 PM
\item 16:  4 PM
\item 17:  5 PM
\item 18:  6 PM
\item 19:  7 PM
\item 20:  8 PM
\item 21:  9 PM
\item 22: 10 PM
\item 23: 11 PM
}
\item \code{holiday}:
\itemize{
\item Whether the day is a holiday or not according to the \href{http://dchr.dc.gov/page/holiday-schedule}{Human Resources page of DC}.
\item 0: No
\item 1: Yes
}
\item \code{weekday}:
\itemize{
\item The day of a week
\item 0: Sunday
\item 1: Monday
\item 2: Tuesday
\item 3: Wednesday
\item 4: Thursday
\item 5: Friday
\item 6: Saturday
}
\item \code{workingday}:
\itemize{
\item Whether the day is a workday (Monday - Friday)
\item 0: No
\item 1: Yes
}
\item \code{weathersit}:
\itemize{
\item 1: Clear, Few clouds, Partly cloudy, Partly cloudy
\item 2: Mist + Cloudy, Mist + Broken clouds, Mist + Few clouds, Mist
\item 3: Light Snow, Light Rain + Thunderstorm + Scattered clouds, Light Rain + Scattered clouds
\item 4: Heavy Rain + Ice Pallets + Thunderstorm + Mist, Snow + Fog
}
\item \code{temp}:
\itemize{
\item Normalized temperature in Celsius.
\item The values are derived via \eqn{\frac{(t-t_{min})}{(t_{max}-t_{min})}}{(t-t[min])/(t[max]-t[min])}, t_min=-8, t_max=+39
}
\item \code{atemp}:
\itemize{
\item Normalized feeling temperature in Celsius.
\item The values are derived via \eqn{\frac{(t-t_{min})}{(t_{max}-t_{min})}}{(t-t[min])/(t[max]-t[min])}, t_min=-16, t_max=+50
}
\item \code{hum}:
\itemize{
\item Normalized humidity.
\item The values are divided to 100 (max)
}
\item \code{windspeed}:
\itemize{
\item Normalized wind speed.
\item The values are divided to 67 (max)
}
\item \code{casual}:
\itemize{
\item Count of casual users
}
\item \code{registered}:
\itemize{
\item Count of registered users
}
\item \code{cnt}:
\itemize{
\item Count of total rental bikes including both casual and registered
}
}
}
\source{
Hadi Fanaee-T

Laboratory of Artificial Intelligence and Decision Support (LIAAD), University of Porto

INESC Porto, Campus da FEUP

Rua Dr. Roberto Frias, 378

4200 - 465 Porto, Portugal
}
\usage{
bike_sharing_daily
}
\description{
Bike sharing systems are new generation of traditional bike rentals where
whole process from membership, rental and return back has become automatic.
Through these systems, user is able to easily rent a bike from a particular
position and return back at another position. Currently, there are about
over 500 bike-sharing programs around the world which is composed of over
500 thousands bicycles. Today, there exists great interest in these systems
due to their important role in traffic, environmental and health issues.
}
\details{
Apart from interesting real world applications of bike sharing systems, the
characteristics of data being generated by these systems make them attractive
for the research. Opposed to other transport services such as bus or subway,
the duration of travel, departure and arrival position is explicitly recorded
in these systems. This feature turns bike sharing system into a virtual
sensor network that can be used for sensing mobility in the city. Hence, it
is expected that most of important events in the city could be detected via
monitoring these data.
}
\references{
Original Source: \url{http://capitalbikeshare.com/system-data}

Weather Information: \url{http://www.freemeteo.com}

Holiday Schedule: \url{http://dchr.dc.gov/page/holiday-schedule}
}
\keyword{datasets}
