% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridges_docs.R
\docType{data}
\name{bridges}
\alias{bridges}
\title{Pittsburgh Bridges Data Set}
\format{A data frame with 108 observations on the following 13 variables.
\itemize{
\item \code{identif}
\itemize{
\item identifier of the examples
}
\item \code{river}
\itemize{
\item A, M, O, Y
}
\item \code{location}
\itemize{
\item Location of Bridge
}
\item \code{erected}
\itemize{
\item Year built
}
\item \code{purpose}
\itemize{
\item WALK, AQUEDUCT, RR, HIGHWAY
}
\item \code{length}
\itemize{
\item 804 - 4558
}
\item \code{lanes}
\itemize{
\item 1, 2, 4, 6
}
\item \code{clear_g}
\itemize{
\item N, G
}
\item \code{t_or_d}
\itemize{
\item THROUGH, DECK
}
\item \code{material}
\itemize{
\item WOOD, IRON, STEEL
}
\item \code{span}
\itemize{
\item SHORT, MEDUIM, LONG
}
\item \code{rel_l}
\itemize{
\item S, S-F, F
}
\item \code{type}
\itemize{
\item WOOD, SUSPEN, SIMPLE-T, ARCH, CANTILEV, CONT-T
}
}}
\source{
Yoram Reich & Steven J. Fenves
Department of Civil Engineering
and
Engineering Design Research Center
Carnegie Mellon University
Pittsburgh, PA 15213
}
\usage{
bridges
}
\description{
Data containing examples of Pittsburgh bridges and the relevant surrounding
area.
}
\details{
This data set is non-discretized, meaning the numeric properties
were left intact.
}
\references{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/bridges/bridges.data.version1}
\url{https://archive.ics.uci.edu/ml/datasets/Pittsburgh+Bridges}
}
\keyword{datasets}
