% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_eval_docs.R
\docType{data}
\name{car_eval}
\alias{car_eval}
\title{Car Evaluation Data Set}
\format{A data frame with 1728 observations on the following 7 variables.
\itemize{
\item \code{buying}
\itemize{
\item vhigh, high, med, low.
}
\item \code{maint}
\itemize{
\item vhigh, high, med, low.
}
\item \code{doors}
\itemize{
\item 2, 3, 4, 5more.
}
\item \code{persons}
\itemize{
\item 2, 4, more.
}
\item \code{lug_boot}
\itemize{
\item small, med, big.
}
\item \code{safety}
\itemize{
\item low, med, high.
}
\item \code{class_value}
\itemize{
\item unacc, acc, good, vgood
}
}}
\source{
Marko Bohanec (marko.bohanec '@' ijs.si)
Blaz Zupan (blaz.zupan '@' ijs.si)
}
\usage{
car_eval
}
\description{
Car Evaluation Database was derived from a simple hierarchical decision model
originally developed for the demonstration of DEX.
}
\references{
M. Bohanec, V. Rajkovic: Expert system for decision making. Sistemica 1(1), pp. 145-157, 1990.)
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/car/car.data}
\url{https://archive.ics.uci.edu/ml/datasets/Car+Evaluation}
}
\keyword{datasets}
