% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_fires_docs.R
\docType{data}
\name{forest_fires}
\alias{forest_fires}
\title{Forest Fires Data Set}
\format{A data frame with 517 observations on the following 13 variables.
\itemize{
\item \code{X}
\itemize{
\item x-axis spatial coordinate within the Montesinho park map: 1 to 9
}
\item \code{Y}
\itemize{
\item y-axis spatial coordinate within the Montesinho park map: 2 to 9
}
\item \code{month}
\itemize{
\item month of the year: "jan" to "dec"
}
\item \code{day}
\itemize{
\item day of the week: "mon" to "sun"
}
\item \code{FFMC}
\itemize{
\item FFMC index from the FWI system: 18.7 to 96.20
}
\item \code{DMC}
\itemize{
\item DMC index from the FWI system: 1.1 to 291.3
}
\item \code{DC}
\itemize{
\item DC index from the FWI system: 7.9 to 860.6
}
\item \code{ISI}
\itemize{
\item ISI index from the FWI system: 0.0 to 56.10
}
\item \code{temp}
\itemize{
\item temperature in Celsius degrees: 2.2 to 33.30
}
\item \code{RH}
\itemize{
\item relative humidity in %: 15.0 to 100
}
\item \code{wind}
\itemize{
\item wind speed in km/h: 0.40 to 9.40
}
\item \code{rain}
\itemize{
\item outside rain in mm/m2 : 0.0 to 6.4
}
\item \code{area}
\itemize{
\item the burned area of the forest (in ha): 0.00 to 1090.84#'
}
}}
\source{
Paulo Cortez, pcortez '@' dsi.uminho.pt, Department of Information Systems, University of Minho, Portugal.
Aníbal Morais, araimorais '@' gmail.com, Department of Information Systems, University of Minho, Portugal.
}
\usage{
forest_fires
}
\description{
The aim is to predict the burned area of forest fires, in the northeast
region of Portugal, by using meteorological and other data
}
\references{
\href{http://www.dsi.uminho.pt/~pcortez/fires.pdf}{ P. Cortez and A. Morais. A Data Mining Approach to Predict Forest Fires using Meteorological Data. In J. Neves, M. F. Santos and J. Machado Eds., New Trends in Artificial Intelligence, Proceedings of the 13th EPIA 2007 - Portuguese Conference on Artificial Intelligence, December, Guimarães, Portugal, pp. 512-523, 2007. APPIA, ISBN-13 978-989-95618-0-9}
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/forest-fires/forestfires.csv}
\url{https://archive.ics.uci.edu/ml/datasets/Forest+Fires}
}
\keyword{datasets}
