% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass_docs.R
\docType{data}
\name{glass}
\alias{glass}
\title{Glass Identification Data Set}
\format{A data frame with 214 observations on the following 11 variables.
\itemize{
\item \code{ID}: 1 to 214
\item \code{RI}: refractive index
\item \code{Na}: weight percent in corresponding oxide, as are attributes 4-10)
\item \code{Mg}: Magnesium
\item \code{Al}: Aluminum
\item \code{Si}: Silicon
\item \code{K}: Potassium
\item \code{Ca}: Calcium
\item \code{Ba}: Barium
\item \code{Fe}: Iron
\item \code{Type}: Class attribute
\itemize{
\item 1: building_windows_float_processed
\item 2: building_windows_non_float_processed
\item 3: vehicle_windows_float_processed
\item 4: vehicle_windows_non_float_processed (none in this database)
\item 5: containers
\item 6: tableware
\item 7: headlamps
}
}}
\source{
B. German
Central Research Establishment
Home Office Forensic Science Service
Aldermaston, Reading, Berkshire RG7 4PN

Vina Spiehler, Ph.D., DABFT
Diagnostic Products Corporation
(213) 776-0180 (ext 3014)
}
\usage{
glass
}
\description{
The study of classification of types of glass was motivated by criminological
investigation. At the scene of the crime, the glass left can be used as
evidence...if it is correctly identified!
}
\references{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/glass/glass.data}
\url{https://archive.ics.uci.edu/ml/datasets/Glass+Identification}
}
\keyword{datasets}
