% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart_disease_processed_docs.R
\docType{data}
\name{heart_disease_cl}
\alias{heart_disease_cl}
\alias{heart_disease_hu}
\alias{heart_disease_va}
\alias{heart_disease_ch}
\title{Heart Disease Processed Data Sets}
\format{
Four \code{data.frames} with a varying number of observations that contain
the following 14 variables.
\itemize{
\item \code{age}: age in years
\item \code{sex}: sex (1 = male; 0 = female)
\item \code{cp}: chest pain type
\itemize{
\item Value 1: typical angina
\item Value 2: atypical angina
\item Value 3: non-anginal pain
\item Value 4: asymptomatic
}
\item \code{trestbps}: resting blood pressure (in mm Hg on admission to the hospital)
\item \code{chol}: serum cholestoral in mg/dl
\item \code{fbs}: fasting blood sugar > 120 mg/dl (1 = true; 0 = false)
\item \code{restecg}: resting electrocardiographic results
\itemize{
\item Value 0: normal
\item Value 1: having ST-T wave abnormality (T wave inversions and/or ST elevation or depression of > 0.05 mV)
\item Value 2: showing probable or definite left ventricular hypertrophy by Estes' criteria
}
\item \code{thalach}: maximum heart rate achieved
\item \code{exang}: exercise induced angina (1 = yes; 0 = no)
\item \code{oldpeak}: ST depression induced by exercise relative to rest
\item \code{slope}: the slope of the peak exercise ST segment
\itemize{
\item Value 1: upsloping
\item Value 2: flat
\item Value 3: downsloping
}
\item \code{ca}: number of major vessels (0-3) colored by flourosopy
\item \code{thal}: See below
\itemize{
\item Value 3: normal
\item Value 6: fixed defect
\item Value 7: reversable defect
}
\item \code{num}: diagnosis of heart disease (angiographic disease status)
\itemize{
\item Value 0: < 50\% diameter narrowing
\item Value 1: > 50\% diameter narrowing
}
}

An object of class \code{data.frame} with 294 rows and 14 columns.

An object of class \code{data.frame} with 200 rows and 14 columns.

An object of class \code{data.frame} with 123 rows and 14 columns.
}
\source{
\enumerate{
\item Hungarian Institute of Cardiology. Budapest: Andras Janosi, M.D.
\item University Hospital, Zurich, Switzerland: William Steinbrunn, M.D.
\item University Hospital, Basel, Switzerland: Matthias Pfisterer, M.D.
\item V.A. Medical Center, Long Beach and Cleveland Clinic Foundation: Robert Detrano, M.D., Ph.D.
}
}
\usage{
heart_disease_cl

heart_disease_hu

heart_disease_va

heart_disease_ch
}
\description{
Detecting the presence of heart disease in patients.
}
\details{
The data was collected from the four following locations:
\enumerate{
\item \code{heart_disease_cl}: Cleveland Clinic Foundation
\item \code{heart_disease_hu}: Hungarian Institute of Cardiology, Budapest
\item \code{heart_disease_va}: V.A. Medical Center, Long Beach, CA
\item \code{heart_disease_ch}: University Hospital, Zurich, Switzerland
}

\tabular{rrr}{
Database        \tab Instances  \cr
Cleveland     \tab 303        \cr
Hungarian     \tab 294        \cr
Switzerland   \tab 123        \cr
Long Beach VA \tab 200        \cr
}
}
\references{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/heart-disease/processed.cleveland.data}
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/heart-disease/processed.hungarian.data}
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/heart-disease/processed.switzerland.data}
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/heart-disease/processed.va.data}
\url{https://archive.ics.uci.edu/ml/datasets/Heart+Disease}
}
\keyword{datasets}
