% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepatitis_docs.R
\docType{data}
\name{hepatitis}
\alias{hepatitis}
\title{Hepatitis Data Set}
\format{A data frame with 6497 observations (1599 Red and 4898 White) on the following 12 variables.
\itemize{
\item \code{class}
\itemize{
\item Die or Live
}
\item \code{age}
\itemize{
\item Integer
}
\item \code{sex}
\itemize{
\item Male, Female
}
\item \code{steroid}
\itemize{
\item No, Yes
}
\item \code{antivirals}
\itemize{
\item No, Yes
}
\item \code{fatigue}
\itemize{
\item No, Yes
}
\item \code{malaise}
\itemize{
\item No, Yes
}
\item \code{anorexia}
\itemize{
\item No, Yes
}
\item \code{liver_big}
\itemize{
\item No, Yes
}
\item \code{liver_firm}
\itemize{
\item No, Yes
}
\item \code{spleen_palpable}
\itemize{
\item No, Yes
}
\item \code{spiders}
\itemize{
\item No, Yes
}
\item \code{ascites}
\itemize{
\item No, Yes
}
\item \code{varices}
\itemize{
\item No, Yes
}
\item \code{bilirubin}
\itemize{
\item Numeric
\item This can also be treated as a factor
}
\item \code{alk_phosphate}
\itemize{
\item Integer
}
\item \code{sgot}
\itemize{
\item Integer
}
\item \code{albumin}
\itemize{
\item Numeric
}
\item \code{protime}
\itemize{
\item Integer
}
\item \code{histology}
\itemize{
\item No, Yes
}
}}
\source{
G.Gong (Carnegie-Mellon University) via
Bojan Cestnik
Jozef Stefan Institute
Jamova 39
61000 Ljubljana
Yugoslavia (tel.: (38)(+61) 214-399 ext.287)
}
\usage{
hepatitis
}
\description{
This data set contains information on folks that suffer from hepatitis.
}
\references{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/hepatitis/hepatitis.data}
\url{https://archive.ics.uci.edu/ml/datasets/hepatitis}
}
\keyword{datasets}
