% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucidata.R
\docType{package}
\name{ucidata-package}
\alias{ucidata}
\alias{ucidata-package}
\title{ucidata: Collection of Datasets from the UC Irvine Machine Learning Repository}
\description{
In this package, there are select datasets from the UC Irvine
Machine Learning Repository. To be included within this package, the
data sets had to conform to being reasonable in size while also allowing
for a wide variety of visualizations and models to be formed.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
\keyword{internal}
