% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_datasets.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine Data Set}
\format{A data frame with 6497 observations (1599 Red and 4898 White) on the following 12 variables.
\itemize{
\item fixed acidity
\item volatile acidity
\item citric acid
\item residual sugar
\item chlorides
\item free sulfur dioxide
\item total sulfur dioxide
\item density
\item pH
\item sulphates
\item alcohol
\item quality
\itemize{
\item Score between 0 and 10 based on sensor reading
}
\item color
\itemize{
\item \code{"White"} or \code{"Red"}
}
}}
\source{
P. Cortez, A. Cerdeira, F. Almeida, T. Matos and J. Reis.
Modeling wine preferences by data mining from physicochemical properties.
In Decision Support Systems, Elsevier, 47(4):547-553. ISSN: 0167-9236.
}
\usage{
wine
}
\description{
This data set is the combination of two datasets that were created, using red and white wine samples.
The inputs include objective tests (e.g. PH values) and the output is based on sensory data
(median of at least 3 evaluations made by wine experts). Each expert graded the wine quality
between 0 (very bad) and 10 (very excellent). Several data mining methods were applied to model
these datasets under a regression approach. The support vector machine model achieved the
best results. Several metrics were computed: MAD, confusion matrix for a fixed error tolerance (T),
etc. Also, we plot the relative importances of the input variables (as measured by a sensitivity
analysis procedure).
}
\references{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/wine/wine.names}
\url{https://archive.ics.uci.edu/ml/datasets/wine}
}
\keyword{datasets}
