% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucidata.R
\docType{package}
\name{ucidata-package}
\alias{ucidata}
\alias{ucidata-package}
\title{ucidata: Collection of Datasets from the UCI Irving Machine Learning Repository}
\description{
Varying datasets from the UCI Irving Machine Learning Repository
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
\keyword{internal}
