% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucidata.R
\docType{package}
\name{ucidata-package}
\alias{ucidata}
\alias{ucidata-package}
\title{ucidata: Collection of Datasets from the UC Irvine Machine Learning Repository}
\description{
Select datasets from the UC Irvine 
    Machine Learning Repository that conform to being reasonable in 
    size while also allowing for a wide variety of visualizations and models
    to be formed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless/ucidata}
  \item Report bugs at \url{https://github.com/coatless/ucidata/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
\keyword{internal}
